Allow s3 key to be overriden with s3_key_stem
The key stem is the last part of the object name excluding the file
format: e.g. "stem" in s3://some-bucket/stem.json.gz
Prior implementation forced the stem of the key to be the current
timestamp. If s3_key_stem is not provided in configuration, the current
timestamp will be used by default as before. Otherwise, the value for
s3_key_stem will be used.